requirements.txt
fastapi
uvicorn
twilio
gspread
google-auth
python-dotenv
python-multipart
pandas
openpyxl
